chore: move into react folder and expose method
